flaskApp/requirements.txt
flask
flask_cors
pymongo
requests
PyJWT
dnspython
